+++ src/components/collect/SearchedCollection/SearchedCollection.tsx
@@ -1,6 +1,5 @@
 import { useGetMemesFromCollectionByKeyword } from "@/application/hooks";
-import { InfiniteMemeList } from "@/components/meme/InfiniteMemeList";
-import { MemeLongPressContainer } from "@/components/meme/LongPress";
+import { InfiniteMemeList, MemeLongPressContainer } from "@/components/meme";
 
 interface Props {
   searchQuery: string;
