+++ src/components/home/MemeListContainer/MemeList/CommonMemeList.tsx
@@ -1,6 +1,5 @@
 import { useGetMemesBySort } from "@/application/hooks";
-import { InfiniteMemeList } from "@/components/meme/InfiniteMemeList";
-import { MemeLongPressContainer } from "@/components/meme/LongPress";
+import { InfiniteMemeList, MemeLongPressContainer } from "@/components/meme";
 
 import type { MemeListType } from "../type";
