+++ src/components/home/MemeListContainer/MemeList/UserFindMemeList.tsx
@@ -1,6 +1,5 @@
 import { useGetUserFindMemes } from "@/application/hooks";
-import { InfiniteMemeList } from "@/components/meme/InfiniteMemeList";
-import { MemeLongPressContainer } from "@/components/meme/LongPress";
+import { InfiniteMemeList, MemeLongPressContainer } from "@/components/meme";
 
 interface Props {
   userId: number | undefined;
