+++ src/components/meme/MemeInfo/RelativeMemeList.tsx
@@ -1,8 +1,7 @@
 import { css } from "twin.macro";
 
 import { useAuth, useGetMemesBySort, useIntersect } from "@/application/hooks";
-import { Masonry } from "@/components/common/Masonry";
-import { MemeItem } from "@/components/meme/MemeItem";
+import { InfiniteMemeList, MemeLongPressContainer } from "@/components/meme";
 
 export const RelativeMemeList = () => {
   const { data: memeList, fetchNextPage } = useGetMemesBySort("popular");
@@ -24,11 +23,9 @@ export const RelativeMemeList = () => {
           `,
         ]}
       >
-        <Masonry columns={2} spacing={9}>
-          {memeList.map((meme) => (
-            <MemeItem key={meme.memeId} meme={meme} />
-          ))}
-        </Masonry>
+        <MemeLongPressContainer memeList={memeList}>
+          <InfiniteMemeList memeList={memeList} onEndReached={fetchNextPage} />
+        </MemeLongPressContainer>
         <div className="h-20" ref={ref} />
       </div>
     </>
