+++ src/components/meme/index.ts
@@ -1,0 +1,4 @@
+export * from "./InfiniteMemeList";
+export * from "./LongPress";
+export * from "./MemeInfo";
+export * from "./MemeItem";
